refactoring: rm unused interface
