+++ src/test/java/lotto/UserTest.java
@@ -5,12 +5,12 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-class UserImplTest {
+class UserTest {
 
   @Test
   @DisplayName("유저에게 14000원을 주면 14개의 티켓을 구매한다")
   void buyLottoTickets() {
-    User user = new UserImpl(14000, new LottoEventImpl());
+    User user = new User(14000, new LottoEvent());
     user.buyAllLottoTickets();
 
     assertEquals(14, user.getLottoTickets().size());
@@ -21,7 +21,7 @@ class UserImplTest {
   void createLottoResult() {
     FOURTHRankMockLottoEvent mockLottoEvent = new FOURTHRankMockLottoEvent();
 
-    User user = new UserImpl(5000, mockLottoEvent);
+    User user = new User(5000, mockLottoEvent);
     user.buyAllLottoTickets();
 
     assertEquals(user.getLottoTickets().size(), user.evaluateLottoTickets().size());
@@ -30,7 +30,7 @@ class UserImplTest {
   }
 
 
-  private static class FOURTHRankMockLottoEvent implements LottoEvent {
+  private static class FOURTHRankMockLottoEvent extends LottoEvent {
 
     public FOURTHRankMockLottoEvent() {
     }
@@ -43,11 +43,6 @@ class UserImplTest {
     @Override
     public int getLottoTicketPrice() {
       return 1000;
-    }
-
-    @Override
-    public void setWinningLottoTicketNumberSet(LottoTicketNumberSet winningLottoTicketNumberSet) {
-      // do nothing
     }
 
     @Override
